+++ l10n/es_CL.js
@@ -15,8 +15,10 @@ OC.L10N.register(
     "This week" : "Esta semana",
     "Status message" : "Mensaje de estado",
     "Share screen" : "Compartir la pantalla",
+    "Menu" : "Menú",
     "Report a bug" : "Reporta un detalle",
     "Log out" : "Cerrar sesión",
+    "Away" : "Ausente",
     "Invisible" : "Invisible"
 },
 "nplurals=3; plural=n == 1 ? 0 : n != 0 && n % 1000000 == 0 ? 1 : 2;");