+++ l10n/ko.js
@@ -16,10 +16,13 @@ OC.L10N.register(
     "Cancel" : "취소",
     "Share screen" : "화면 공유",
     "Loading …" : "불러오는 중 ...",
-    "View profile" : "프로필 보기",
+    "Menu" : "메뉴",
     "Report a bug" : "버그 보고",
+    "Back" : "뒤로",
+    "View profile" : "프로필 보기",
     "Away" : "자리비움",
     "Do not disturb" : "방해 없음",
-    "Invisible" : "숨겨짐"
+    "Invisible" : "숨겨짐",
+    "Unknown status" : "알 수 없는 상태"
 },
 "nplurals=1; plural=0;");