+++ src/talk/renderer/init.js
@@ -1,27 +1,14 @@
-/*
- * @copyright Copyright (c) 2023 Grigorii Shartsev <me@shgk.me>
- *
- * @author Grigorii Shartsev <me@shgk.me>
- *
- * @license AGPL-3.0-or-later
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+/**
+ * SPDX-FileCopyrightText: 2023 Nextcloud GmbH and Nextcloud contributors
+ * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 import { loadServerCss } from '../../shared/resource.utils.js'
 import { appData } from '../../app/AppData.js'
 import { getCapabilities } from '../../shared/ocs.service.js'
+import { setInitialState } from '../../shared/initialState.service.js'
+import { subscribe } from '@nextcloud/event-bus'
+import { getCurrentUser } from '@nextcloud/auth'
 
 /**
  * @return {Promise<void>}
@@ -170,5 +157,23 @@ export function initTalkHashIntegration(talkInstance) {
 			// TODO: make soft restart?
 			window.TALK_DESKTOP.relaunch()
 		},
+	})
+}
+
+/**
+ * Enable or disable the sound for notifications and calls on DND user status change
+ */
+export function initPlaySoundManagementOnUserStatus() {
+	subscribe('user_status:status.updated', (userStatus) => {
+		if (userStatus.userId !== getCurrentUser().uid) {
+			return
+		}
+		// TODO: add setting to define the default value for playSound
+		const playSoundDefault = true
+		// Disable if DND
+		const playSound = userStatus.status === 'dnd' ? false : playSoundDefault
+		// Notification's sound in the Notifications app is controlled via initial state only
+		setInitialState('notifications', 'sound_notification', playSound)
+		setInitialState('notifications', 'sound_talk', playSound)
 	})
 }