+++ src/talk/talk.window.js
@@ -1,25 +1,9 @@
-/*
- * @copyright Copyright (c) 2022 Grigorii Shartsev <me@shgk.me>
- *
- * @author Grigorii Shartsev <me@shgk.me>
- *
- * @license GNU AGPL version 3 or any later version
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+/**
+ * SPDX-FileCopyrightText: 2022 Nextcloud GmbH and Nextcloud contributors
+ * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-const { BrowserWindow } = require('electron')
+const { BrowserWindow, screen } = require('electron')
 const { applyExternalLinkHandler } = require('../app/externalLinkHandlers.js')
 const { applyContextMenu } = require('../app/applyContextMenu.js')
 const { applyDownloadNotification } = require('../app/applyDownloadNotification.js')
@@ -31,6 +15,9 @@ const { getBrowserWindowIcon } = require('../shared/icons.utils.js')
  * @return {import('electron').BrowserWindow}
  */
 function createTalkWindow() {
+	const primaryDisplay = screen.getPrimaryDisplay()
+	const { width: screenWidth, height: screenHeight } = primaryDisplay.workAreaSize
+
 	const talkWindowOptions = {
 		minWidth: 600,
 		minHeight: 400,
@@ -55,8 +42,8 @@ function createTalkWindow() {
 
 	const window = new BrowserWindow({
 		...talkWindowOptions,
-		width: 1280,
-		height: 800,
+		width: Math.min(1680, screenWidth),
+		height: Math.min(1050, screenHeight),
 		show: false,
 	})
 
@@ -69,8 +56,8 @@ function createTalkWindow() {
 
 	applyExternalLinkHandler(window, {
 		...talkWindowOptions,
-		width: 800,
-		height: 600,
+		width: Math.min(800, screenWidth),
+		height: Math.min(600, screenHeight),
 	})
 
 	applyContextMenu(window)