+++ webpack.renderer.config.js
@@ -1,22 +1,6 @@
-/*
- * @copyright Copyright (c) 2022 Grigorii Shartsev <grigorii.shartsev@nextcloud.com>
- *
- * @author Grigorii Shartsev <grigorii.shartsev@nextcloud.com>
- *
- * @license GNU AGPL version 3 or any later version
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
+/**
+ * SPDX-FileCopyrightText: 2022 Nextcloud GmbH and Nextcloud contributors
+ * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
 require('dotenv').config()
@@ -121,11 +105,10 @@ const webpackRendererConfig = mergeWithRules({
 	resolve: {
 		alias: {
 			'@talk': TALK_PATH,
-			...createPatcherAliases('@nextcloud/initial-state'),
-			...createPatcherAliases('@nextcloud/axios'),
+			// To reuse modules between Talk Desktop and Talk, otherwise Talk has its own from its node_modules
+			'@nextcloud/axios': path.resolve(__dirname, 'node_modules', '@nextcloud/axios/dist/index.mjs'),
+			// Patched packages
 			...createPatcherAliases('@nextcloud/router'),
-			...createPatcherAliases('@nextcloud/auth'),
-			...createPatcherAliases('@nextcloud/notify_push'),
 		},
 	},